unify map hook
