+++ applications/fishing-map/src/features/sidebar/context-areas/ContextAreaLayerPanel.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import cx from 'classnames'
 import { UrlDataviewInstance, AsyncReducerStatus } from 'types'
 import { useSelector } from 'react-redux'
-import useClickedOutside from 'hooks/useClickedOutside'
+import useClickedOutside from 'hooks/use-clicked-outside'
 import { formatInfoField } from 'utils/info'
 import { Vessel } from '@globalfishingwatch/api-types'
 import { Switch, IconButton, Tooltip, ColorBar } from '@globalfishingwatch/ui-components'
