+++ applications/fishing-map/src/hooks/use-clicked-outside.ts
